requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
fpdf2==2.7.9
openpyxl==3.1.2
